Convert Oauth, Polling, and Projects to TS (#1306)
