+++ commands/project/listBuilds.ts
@@ -17,9 +17,9 @@ const { uiBetaTag, uiLink } = require('../../lib/ui');
 const { loadAndValidateOptions } = require('../../lib/validation');
 const {
   getProjectConfig,
-  getProjectDetailUrl,
   validateProjectConfig,
 } = require('../../lib/projects');
+const { getProjectDetailUrl } = require('../../lib/projects/urls');
 const moment = require('moment');
 const { promptUser } = require('../../lib/prompts/promptUtils');
 const { isHubSpotHttpError } = require('@hubspot/local-dev-lib/errors/index');
