+++ lib/constants.ts
@@ -61,12 +61,12 @@ export const PROJECT_ERROR_TYPES = {
   SUBDEPLOY_FAILED: 'DeployPipelineErrorType.DEPENDENT_SUBDEPLOY_FAILED',
 } as const;
 
-export const PROJECT_TASK_TYPES = {
+export const PROJECT_TASK_TYPES: { [key: string]: string } = {
   PRIVATE_APP: 'private app',
   PUBLIC_APP: 'public app',
   APP_FUNCTION: 'function',
   CRM_CARD_V2: 'card',
-} as const;
+};
 
 export const PROJECT_COMPONENT_TYPES = {
   PROJECTS: 'projects',
