+++ lib/localDev.ts
@@ -38,10 +38,8 @@ const {
   isAppDeveloperAccount,
   isDeveloperTestAccount,
 } = require('./accountTypes');
-const {
-  handleProjectUpload,
-  pollProjectBuildAndDeploy,
-} = require('./projects');
+const { handleProjectUpload } = require('./projects/upload');
+const { pollProjectBuildAndDeploy } = require('./projects/buildAndDeploy');
 const {
   PROJECT_ERROR_TYPES,
   PROJECT_BUILD_TEXT,
@@ -248,9 +246,8 @@ const createDeveloperTestAccountForLocalDev = async (
   let currentPortalCount = 0;
   let maxTestPortals = 10;
   try {
-    const validateResult = await validateDevTestAccountUsageLimits(
-      accountConfig
-    );
+    const validateResult =
+      await validateDevTestAccountUsageLimits(accountConfig);
     if (validateResult) {
       currentPortalCount = validateResult.results
         ? validateResult.results.length
@@ -306,9 +303,8 @@ const createDeveloperTestAccountForLocalDev = async (
 
 // Prompt user to confirm usage of an existing developer test account that is not currently in the config
 const useExistingDevTestAccount = async (env, account) => {
-  const useExistingDevTestAcct = await confirmUseExistingDeveloperTestAccountPrompt(
-    account
-  );
+  const useExistingDevTestAcct =
+    await confirmUseExistingDeveloperTestAccountPrompt(account);
   if (!useExistingDevTestAcct) {
     logger.log('');
     logger.log(
